format spring-3.1.x-scenario
